# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: ManufactureOS ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchaseorders:
    get:
      operationId: listPurchaseOrders
      tags: [purchaseorders]
      summary: List purchase orders
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: vendorId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Purchase orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrdersListResponse"
    post:
      operationId: createPurchaseOrder
      tags: [purchaseorders]
      summary: Create purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseOrderRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchaseorders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchaseorders]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
    put:
      operationId: updatePurchaseOrder
      tags: [purchaseorders]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePurchaseOrderRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchaseorders/{id}/receive:
    post:
      operationId: receivePurchaseOrder
      tags: [purchaseorders]
      summary: Receive against purchase order
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceivePORequest"
      responses:
        "200":
          description: Receipt posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # WORK ORDERS
...
  /quality/inspections:
    get:
      operationId: listInspections
      tags: [quality]
      summary: List inspections
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Inspections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectionsListResponse"
    post:
      operationId: createInspection
      tags: [quality]
      summary: Create inspection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInspectionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
  /quality/nonconformances:
    get:
      operationId: listNonconformances
      tags: [quality]
      summary: List nonconformances
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Nonconformances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NonconformancesListResponse"
    post:
      operationId: createNonconformance
      tags: [quality]
      summary: Create nonconformance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNonconformanceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Nonconformance"

  # MRP
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 25
...
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - total
        - totalPages

    # USER / AUTH
...
    PurchaseOrderLine:
      type: object
      properties:
        id:
          type: string
        lineNumber:
          type: integer
        itemId:
          type: string
        itemNumber:
          type: string
        itemName:
          type: string
        quantity:
          type: number
        quantityReceived:
          type: number
        uom:
          type: string
        unitCost:
          type: number
        lineTotal:
          type: number
        requestedDate:
          type: string
          format: date
        promisedDate:
          type: string
          format: date
        notes:
          type: string
      required:
        - id
        - lineNumber
        - itemId
        - quantity
        - unitCost
    PurchaseOrder:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        vendorId:
          type: string
        vendorName:
          type: string
        status:
          type: string
          enum: [draft, sent, acknowledged, partially_received, received, cancelled, closed]
        orderDate:
          type: string
          format: date
        requestedDate:
          type: string
          format: date
        promisedDate:
          type: string
          format: date
        lines:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderLine"
        subtotal:
          type: number
        taxAmount:
          type: number
        totalAmount:
          type: number
        warehouseId:
          type: string
        notes:
          type: string
        terms:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - number
        - vendorId
        - status
    CreatePurchaseOrderRequest:
      type: object
      properties:
        vendorId:
          type: string
        orderDate:
          type: string
          format: date
        requestedDate:
          type: string
          format: date
        warehouseId:
          type: string
        lines:
          type: array
          items:
            type: object
            properties:
              itemId:
                type: string
              quantity:
                type: number
              unitCost:
                type: number
              requestedDate:
                type: string
                format: date
              notes:
                type: string
            required:
              - itemId
              - quantity
              - unitCost
        notes:
          type: string
        terms:
          type: string
      required:
        - vendorId
        - lines
    UpdatePurchaseOrderRequest:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, acknowledged, cancelled]
        requestedDate:
          type: string
          format: date
        promisedDate:
          type: string
          format: date
        warehouseId:
          type: string
        notes:
          type: string
    ReceivePORequest:
      type: object
      properties:
        lines:
          type: array
          items:
            type: object
            properties:
              lineId:
                type: string
              quantityReceived:
                type: number
              lotNumber:
                type: string
              notes:
                type: string
            required:
              - lineId
              - quantityReceived
        receivedDate:
          type: string
          format: date
        notes:
          type: string
      required:
        - lines
    PurchaseOrdersListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrder"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta

    # WORK ORDER
...
    Inspection:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        type:
          type: string
          enum: [incoming, in_process, final]
        status:
          type: string
          enum: [pending, in_progress, passed, failed, conditionally_passed]
        itemId:
          type: string
        itemNumber:
          type: string
        itemName:
          type: string
        quantity:
          type: number
        quantityPassed:
          type: number
        quantityFailed:
          type: number
        reference:
          type: string
        lotNumber:
          type: string
        inspectedBy:
          type: string
        inspectedAt:
          type: string
          format: date-time
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - number
        - type
        - status
        - itemId
    CreateInspectionRequest:
      type: object
      properties:
        type:
          type: string
          enum: [incoming, in_process, final]
        itemId:
          type: string
        quantity:
          type: number
        reference:
          type: string
        lotNumber:
          type: string
        notes:
          type: string
      required:
        - type
        - itemId
        - quantity
    InspectionsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Inspection"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    Nonconformance:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, under_review, pending_disposition, closed, voided]
        severity:
          type: string
          enum: [minor, major, critical]
        itemId:
          type: string
        itemNumber:
          type: string
        defectCode:
          type: string
        disposition:
          type: string
          enum: [use_as_is, rework, scrap, return_to_vendor]
        quantityAffected:
          type: number
        lotNumber:
          type: string
        containmentAction:
          type: string
        rootCause:
          type: string
        correctiveAction:
          type: string
        reportedBy:
          type: string
        assignedTo:
          type: string
        closedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - number
        - title
        - status
        - severity
    CreateNonconformanceRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [minor, major, critical]
        itemId:
          type: string
        defectCode:
          type: string
        quantityAffected:
          type: number
        lotNumber:
          type: string
        containmentAction:
          type: string
        rootCause:
          type: string
        assignedTo:
          type: string
      required:
        - title
        - severity
    NonconformancesListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Nonconformance"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta

    # MRP
...