# lib/api-zod/src/generated/types/createInspectionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ManufactureOS ERP API
 * OpenAPI spec version: 1.0.0
 */
import type { CreateInspectionRequestType } from "./createInspectionRequestType";

export interface CreateInspectionRequest {
  type: CreateInspectionRequestType;
  itemId: string;
  quantity: number;
  reference?: string;
  lotNumber?: string;
  notes?: string;
}